psycopg2_copy/pyproject.toml
[project]
name = "etl-psycopg2-copy"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary",
]
